third_party/WebKit/LayoutTests/media/track/track-text-src-change.html

<!DOCTYPE html>
<title>HTMLTrackElement 'src' attribute mutations</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<video>
    <track src="captions-webvtt/tc013-settings.vtt" default>
</video>
<script>
async_test(function(t) {
    var cues = null;
    var testTrack = document.querySelector("track");
    var stage = 0;
    testTrack.onload = t.step_func(function() {
        switch (stage) {
            case 0:
                cues = testTrack.track.cues;
                assert_equals(cues.length, 4);
                ++stage;
                testTrack.src = "captions-webvtt/tc022-entities.vtt";
                break;
            case 1:
                assert_equals(cues, testTrack.track.cues);
                assert_equals(cues.length, 7);
                assert_equals(cues[cues.length-1].text, 'This & is parsed to the same as &amp;.');
                ++stage;
                testTrack.src = "captions-webvtt/tc013-settings.vtt";
                break;
            case 2:
                assert_equals(cues, testTrack.track.cues);
                assert_equals(cues.length, 4);
                assert_equals(cues[cues.length-1].text, 'I said Bear is coming now!!!! Tab separators.');
                ++stage;
                testTrack.src = "captions-webvtt/tc013-settings.vtt";
                break;
            case 3:
                assert_equals(cues, testTrack.track.cues);
                assert_equals(cues.length, 4);
                assert_equals(cues[cues.length-1].text, 'I said Bear is coming now!!!! Tab separators.');
                t.name = document.title;
                t.done();
                break;
        }
    });
});
</script>
